requirements.txt
backcall==0.2.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
decorator==4.4.2
Django==2.2
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-js-asset==1.2.2
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.8
pycparser==2.20
Pygments==2.7.2
pytz==2020.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5